Use proxy pattern for task queue-related logic
