+++ src/test/java/teammates/test/MockTaskQueuer.java
@@ -19,16 +19,11 @@ public class MockTaskQueuer extends TaskQueuer {
     private List<TaskWrapper> tasksAdded = new ArrayList<>();
 
     @Override
-    protected void addTask(String queueName, String workerUrl, Map<String, String> paramMap, Object requestBody) {
-        TaskWrapper task = new TaskWrapper(queueName, workerUrl, paramMap, requestBody);
-        tasksAdded.add(task);
-    }
-
-    @Override
     protected void addDeferredTask(String queueName, String workerUrl, Map<String, String> paramMap, Object requestBody,
                                    long countdownTime) {
-        // countdown time not tested, thus fallback to another method
-        addTask(queueName, workerUrl, paramMap, requestBody);
+        // countdown time not tested
+        TaskWrapper task = new TaskWrapper(queueName, workerUrl, paramMap, requestBody);
+        tasksAdded.add(task);
     }
 
     @Override
